Address warnings and comments from review

+++ libraries/animation/src/AnimSkeleton.cpp
@@ -34,7 +34,7 @@ AnimSkeleton::AnimSkeleton(const HFMModel& hfmModel) {
         const auto& defor = hfmModel.skinDeformers[i];
         std::vector<HFMCluster> dummyClustersList;
 
-        for (int j = 0; j < defor.clusters.size(); j++) {
+        for (int j = 0; j < (uint32_t) defor.clusters.size(); j++) {
             // cast into a non-const reference, so we can mutate the FBXCluster
             HFMCluster& cluster = const_cast<HFMCluster&>(defor.clusters.at(j));
